requirements.txt
python-bidi
arabic_reshaper
kivy[base]
https://github.com/kivymd/KivyMD/archive/master.zip
kivyir
fastapi[standard]